chore: rearranged functions and files to make it more readable and concise
